refactor: move handle methods out of gateway calss
