+++ src/Framework/PaymentGateways/PaymentGatewayProcessor.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Give\Framework\PaymentGateways;
+
+class PaymentGatewayProcessor {
+
+}
